requirements.txt
contourpy==1.1.0
customtkinter==5.2.0
cycler==0.11.0
darkdetect==0.8.0
et-xmlfile==1.1.0
fonttools==4.42.1
kiwisolver==1.4.5
matplotlib==3.7.3
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
plotly==5.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tenacity==8.2.3
tzdata==2023.3
